shifted add provider to the providerList screen. I'm gonna be honest, all of the account/provider screens need some overhaul. They're all a mess with untyped variables, inconsistent indentation, etc.
